slack-agent/requirements.txt
strands-agents
slack-bolt
squareup
squareup_legacy
python-dotenv
psycopg2-binary